+++ src/cli/test/download-artifacts.js
@@ -1,28 +1,11 @@
-const https = require('https')
-const fs = require('fs')
 const path = require('path')
-const zlib = require('zlib')
 
 const listr = require('listr')
 
+const download = require('../../utils/download')
+const mkdirp = require('../../utils/mkdirp')
 const { fetchArtifacts } = require('../../api/test')
-
 const { humaniseError } = require('../../utils/api-error')
-
-const download = (url, destination) => {
-  return new Promise(resolve => {
-    const file = fs.createWriteStream(destination)
-    https.get(url, response => {
-      if (response.headers['content-encoding'] == 'gzip') {
-        response.pipe(zlib.createGunzip()).pipe(file)
-      } else {
-        response.pipe(file)
-      }
-
-      file.on('finish', () => file.close(resolve))
-    })
-  })
-}
 
 const main = async args => {
   try {
@@ -31,16 +14,7 @@ const main = async args => {
     if (args.json) return console.log(JSON.stringify(response, null, 2))
 
     const directories = [process.cwd(), 'test-artifacts', args.uuid]
-
-    directories.reduce((reducePath, dir) => {
-      reducePath = path.join(reducePath, dir)
-      if (!fs.existsSync(reducePath)) {
-        fs.mkdirSync(reducePath)
-      }
-      return reducePath
-    }, '')
-
-    const directory = path.join(...directories)
+    const directory = mkdirp(directories)
 
     const tasks = new listr(
       [
